requirements.txt
py-clob-client>=0.16.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
